Update theme image and version in package.json and theme.css
The version in package.json has been updated to 1.2.0. In addition to this, the background image in the Sailor Moon theme has been altered. The change includes using a locally stored image in the assets directory, and slight modifications to the background color have been applied in the theme.css.